baby_wallet_backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-decouple==3.8
Pillow==10.4.0
web3==6.15.1
eth-account==0.11.0
eth-utils==3.0.0
requests==2.31.0
redis==5.0.1
